Added faster allan variance calculation method
Added a file that compares the avar_series() code I wrote against
the adev() code in allantools. They are equivalent, except the
numpy use in avar_series() gives it a ~60x speedup.

This is not yet a replacement for adev() as it doesn't handle
timestamps /sampling rates in the same way.