Adding comparative tests for numpy vs old methods
The idea here is to recreate all the current functions
with _np suffixes, e.g. tdev -> tdev_np, so as to compare
the output between new and old versions. Assuming that the new
versions perform better, eventually old code will be completely
replaced with the new code.

Initial tests see speedups in the range 50-200x, so very promising!