Added fallback in case of bottleneck not installed
At the moment, only the MTIE function uses bottleneck. As such, it
seems that making bottleneck an optional dependency is a better
option than requiring it for everyone.

TODO: Check if this is the best way to do optional dependencies. I've
seen some stuff on 'extras_require' keyword in setup.py, should look
into that further.